Check for the presence of smartparens before calling its functions
slime was not working for me today - it would crash before getting to a repl.  

"error in process filter: Symbol's function definition is void: smartparens-strict-mode"

and if the reference to smartparens-strict-mode in packages.el was commented out, it returned 
"error in process filter: Symbol's function definition is void: turn-off-smartparens-mode"

smartparens does not seem to be included in the common-lisp layer, for some reason.  I am using spacemacs-base with the emacs-lisp and common-lisp layers, and exwm as an additional package.

Adding a check to see whether smartparens is present before calling those functions allows slime to reach a repl, even when smartparens is not installed.